% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inflationfactor.R
\name{dnormmix}
\alias{dnormmix}
\title{density of a k-component normal mixture}
\usage{
dnormmix(x, theta)
}
\arguments{
\item{x}{x like dnorm(x, ...}

\item{theta}{parameters of the mixture proportion, mean and sd}
}
\value{
density of a k-component normal mixture
}
\description{
density of a k-component normal mixture
}
\details{
details follow
}
\author{
mvaniterson
}

