% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inflationfactor.R
\name{gibbssampler}
\alias{gibbssampler}
\title{Gibbs sampler}
\usage{
gibbssampler(y, k = 3, niter = 5000, burnin = 2000, verbose = TRUE,
  plot = FALSE, alpha = 1.28, beta = 0.36 * var(y), lam = c(0, 1, -1),
  tau = c(1000, rep(100, 2)), g = c(90, 5, 5), rho1 = 0.1842,
  rho2 = 3.5)
}
\arguments{
\item{y}{numeric vector containing the observed mixture values}

\item{k}{number of components of the mixture default is 3}

\item{niter}{length of the Gibbs sampling run default 5000}

\item{burnin}{length of the Burn-in default is 2000}

\item{verbose}{print progress information default is TRUE}

\item{plot}{plot traces progress default FALSE}

\item{alpha}{priors for variances inverse-Gamma distribution}

\item{beta}{priors for variances inverse-Gamma distribution}

\item{lam}{priors for means Normal distribution}

\item{tau}{priors for means Normal distribution}

\item{g}{priors for proportions Dirichlet distribution}

\item{rho1}{prior for variances Beta distribution}

\item{rho2}{prior for variances Beta distribution}
}
\value{
posterior samples from the fitted mixture distribution
}
\description{
Gibbs sampler that fits normal mixtures
details follow
TODO improve implementation e.g. S3 to allow easy plotting etc.
}
\examples{
theta <- c(0.9, 1, 0, 3, 1)
y <- rnormmix(5000, theta)
gs <- gibbssampler(y, niter=500, burnin=200)
plotTraces(gs)
plotCI(gs)
points(1, 0.9, pch=16, col=2, cex=2)
theta.hat <- apply(gs, 2, median)
plotnormmix(y, theta.hat, n=100)
}
\author{
mvaniterson
}
\references{
Implementation is based on a version from Zhihui Liu \url{https://macsphere.mcmaster.ca/handle/11375/9368}
}

