% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hcpp.R
\docType{data}
\name{hcpp}
\alias{hcpp}
\title{hcpp example data}
\format{list containing two matrices
\describe{
  \item{Y}{matrix with RNAseq counts 58 x 14573}
  \item{F}{matrix with covariates 58 x 3}
}}
\usage{
data(hcpp)
}
\description{
A simulated RNAseq example dataset on 58 samples and 14573 genes
including a set of three covariates
}
\keyword{datasets}

