% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inflationfactor.R
\name{plotCI}
\alias{plotCI}
\title{plot confidence ellipse}
\usage{
plotCI(gs, theta = c("sigma.1", "p.1"), alphas = c(0.95, 0.9, 0.75), ...)
}
\arguments{
\item{gs}{gibssampler object}

\item{theta}{select parameters to plot}

\item{alphas}{alfa level(s) for the confidence ellipses default 0.95, 0.9 and 0.75.}

\item{...}{arguments to ellipse}
}
\value{
plot
}
\description{
plot normal based confidence ellipse on two parameters
}
\details{
details follow not really generic
}
\author{
mvaniterson
}

