% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inflationfactor.R
\name{plotnormmix}
\alias{plotnormmix}
\title{plot normal mixtures}
\usage{
plotnormmix(data, theta, ...)
}
\arguments{
\item{data}{vector of test statistics}

\item{theta}{parameters describing the mixture components}

\item{...}{arguments passed to hist}
}
\value{
return plot with histogram of the data and mixture and individual components
}
\description{
plot normal mixtures
}
\details{
details follow
}
\author{
mvaniterson
}

