% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inflationfactor.R
\name{plotTraces}
\alias{plotTraces}
\title{plot traces of the samples drawn from the conditional distribtions}
\usage{
plotTraces(gs, theta = rep(NA, ncol(gs)))
}
\arguments{
\item{gs}{gibssampler object}

\item{theta}{optionally select parameters to plot}
}
\value{
trace plot of the different parameters
}
\description{
Show traces of the gibbs sampling
}
\details{
details follow not really generic
}
\author{
mvaniterson
}

