% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_hcp}
\alias{rcpparma_hcp}
\title{Rcpp implementation of the HCP algorithm}
\usage{
rcpparma_hcp(F, Y, k, lambda1, lambda2, lambda3, iter)
}
\arguments{
\item{F}{a matrix nxd of known covariates, where n is the number of
subjects and d is the number of known covariates. *must be standardize
(columns have 0 mean and constant SS).}

\item{Y}{a matrix of nxg of expression data (must be standardized (columns
scaled to have constant SS and mean 0). ** use standardize function to standardize F and Y.}

\item{k}{number of inferred hidden components (k is an integer)}

\item{lambda1}{model parameter 1}

\item{lambda2}{model parameter 1}

\item{lambda3}{model parameter 1}

\item{iter}{(optional) iter: number of iterations (default = 100);}
}
\value{
list
Z: matrix of hidden components, dimensionality: nxk
B: matrix of effects of hidden components, dimensionality: kxg
o: value of objective function on consecutive iterations.
}
\description{
Rcpp implementation of the HCP algorithm
}
\details{
objective:
this function solves the following problem:
argmin_{Z,B,U}   ||Y-Z*B||_2 + lambda1*||Z-F*U||_2 + lambda2*||B||_2 +
lambda_3||U||_2

to use the residual data: Residual = Y - Z*B

note: k>0, lambda1>0, lambda2>0, lambda3>0 must be set by the user based on
the data at hand. one can set these values using cross-validation, by
evaluating the "performance" of the  resulting residual data on a desired
task. typically, if lambda>5, then hidden factors match the known covariates closely.
}
\author{
mvaniterson
}
\references{
\url{http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0068141}
}

