% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhcppdata.R
\docType{data}
\name{rhcppdata}
\alias{rhcppdata}
\title{rhcppdata example data}
\format{list containing two matrices
\describe{
  \item{Y}{matrix with RNAseq counts 58 x 14573}
  \item{F}{matrix with covariates 58 x 3}
}}
\usage{
data(rhcppdata)
}
\description{
A simulated RNAseq example dataset on 58 samples and 14573 genes
including a set of three covariates
}
\keyword{datasets}

