% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inflationfactor.R
\name{rnormmix}
\alias{rnormmix}
\title{sample from a normal mixture}
\usage{
rnormmix(n, theta)
}
\arguments{
\item{n}{size}

\item{theta}{parameters}
}
\value{
n samples from a normal mixture with parameters theta
}
\description{
sample from a normal mixture
}
\details{
details follow
}
\author{
mvaniterson
}

